+++ src/packages/core/property-editor/uis/manifests.ts
@@ -13,7 +13,7 @@ import { manifest as orderDirection } from '../../../property-editors/order-dire
 import { manifest as overlaySize } from '../../../property-editors/overlay-size/manifests.js';
 import { manifest as radioButtonList } from '../../../property-editors/radio-button-list/manifests.js';
 import { manifest as select } from '../../../property-editors/select/manifests.js';
-import { manifest as slider } from './slider/manifests.js';
+import { manifest as slider } from '../../../property-editors/slider/manifests.js';
 import { manifest as textArea } from './textarea/manifests.js';
 import { manifest as toggle } from './toggle/manifests.js';
 import { manifest as uploadField } from './upload-field/manifests.js';
