+++ src/packages/property-editors/slider/manifests.ts
