+++ src/packages/property-editors/slider/property-editor-ui-slider.test.ts
